fix some add dto add greenhouse service

+++ src/main/java/org/greenhouse/repository/greenhouse/ConfigurationsRepository.java
@@ -3,4 +3,4 @@ package org.greenhouse.repository.greenhouse;
 import org.greenhouse.entity.greenhouse.Configurations;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ConfigurationsRepository extends JpaRepository<Configurations, Integer> {}
+public interface ConfigurationsRepository extends JpaRepository<Configurations, Long> {}
