+++ src/main/java/org/greenhouse/repository/greenhouse/ControlRepository.java
@@ -3,4 +3,4 @@ package org.greenhouse.repository.greenhouse;
 import org.greenhouse.entity.greenhouse.Control;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ControlRepository extends JpaRepository<Control, Integer> {}
+public interface ControlRepository extends JpaRepository<Control, Long> {}
