+++ src/main/java/org/greenhouse/repository/greenhouse/GreenhousesRepository.java
@@ -3,4 +3,4 @@ package org.greenhouse.repository.greenhouse;
 import org.greenhouse.entity.greenhouse.Greenhouses;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GreenhousesRepository extends JpaRepository<Greenhouses, Integer> {}
+public interface GreenhousesRepository extends JpaRepository<Greenhouses, Long> {}
