+++ src/main/java/org/greenhouse/repository/greenhouse/SeedBedsRepository.java
@@ -3,4 +3,4 @@ package org.greenhouse.repository.greenhouse;
 import org.greenhouse.entity.greenhouse.SeedBeds;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SeedBedsRepository extends JpaRepository<SeedBeds, Integer> {}
+public interface SeedBedsRepository extends JpaRepository<SeedBeds, Long> {}
