+++ src/main/java/org/greenhouse/repository/log/ReceiveLogsRepository.java
@@ -3,4 +3,4 @@ package org.greenhouse.repository.log;
 import org.greenhouse.entity.log.ReceiveLogs;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReceiveLogsRepository extends JpaRepository<ReceiveLogs, Integer> {}
+public interface ReceiveLogsRepository extends JpaRepository<ReceiveLogs, Long> {}
