+++ src/main/java/org/greenhouse/repository/log/SendLogsRepository.java
@@ -3,4 +3,4 @@ package org.greenhouse.repository.log;
 import org.greenhouse.entity.log.SendLogs;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SendLogsRepository extends JpaRepository<SendLogs, Integer> {}
+public interface SendLogsRepository extends JpaRepository<SendLogs, Long> {}
