+++ src/main/java/org/greenhouse/repository/log/TopicsRepository.java
@@ -3,4 +3,4 @@ package org.greenhouse.repository.log;
 import org.greenhouse.entity.log.Topics;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TopicsRepository extends JpaRepository<Topics, Integer> {}
+public interface TopicsRepository extends JpaRepository<Topics, Long> {}
