+++ src/main/java/org/greenhouse/repository/log/readings/HumidityRepository.java
@@ -3,4 +3,4 @@ package org.greenhouse.repository.log.readings;
 import org.greenhouse.entity.log.readings.Humidity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HumidityRepository extends JpaRepository<Humidity, Integer> {}
+public interface HumidityRepository extends JpaRepository<Humidity, Long> {}
