+++ src/main/java/org/greenhouse/repository/log/readings/LightRepository.java
@@ -3,4 +3,4 @@ package org.greenhouse.repository.log.readings;
 import org.greenhouse.entity.log.readings.Light;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LightRepository extends JpaRepository<Light, Integer> {}
+public interface LightRepository extends JpaRepository<Light, Long> {}
