+++ src/main/java/org/greenhouse/repository/log/readings/TemperatureRepository.java
@@ -3,4 +3,4 @@ package org.greenhouse.repository.log.readings;
 import org.greenhouse.entity.log.readings.Temperature;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TemperatureRepository extends JpaRepository<Temperature, Integer> {}
+public interface TemperatureRepository extends JpaRepository<Temperature, Long> {}
