+++ src/main/java/org/greenhouse/repository/sensor/SensorsTypeRepository.java
@@ -3,4 +3,4 @@ package org.greenhouse.repository.sensor;
 import org.greenhouse.entity.sensor.SensorType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SensorsTypeRepository extends JpaRepository<SensorType, Integer> {}
+public interface SensorsTypeRepository extends JpaRepository<SensorType, Long> {}
